+++ test/mocks/MockUniswapV2Router.sol
@@ -13,12 +13,12 @@ contract MockUniswapV2Router02 {
     function removeLiquidity(
         address tokenA,
         address tokenB,
-        uint liquidity,
-        uint amountAMin,
-        uint amountBMin,
+        uint256 liquidity,
+        uint256 amountAMin,
+        uint256 amountBMin,
         address to,
-        uint deadline
-    ) external returns (uint amountA, uint amountB) {
+        uint256 deadline
+    ) external returns (uint256 amountA, uint256 amountB) {
         require(deadline >= block.timestamp, "UniswapV2Router: EXPIRED");
         require(amountAOut >= amountAMin, "UniswapV2Router: INSUFFICIENT_A_AMOUNT");
         require(amountBOut >= amountBMin, "UniswapV2Router: INSUFFICIENT_B_AMOUNT");
